sgdproject/requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.2.1
django-heroku==0.3.1
django-safedelete==0.5.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
mysqlclient==1.4.6
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
rest-condition==1.0.3
sqlparse==0.3.1
whitenoise==5.0.1
